Uses the pathdb project. This example shows creating and using index.
You need to have pathdb project checked out locally,
and built/installed in your local maven repository.
You can do this with `mvn install` from the root of the PathDB project
directory. At the time of this writing, you also need to be in the
`in-memory-tree` branch of the PathDB project.